pyproject.toml

[project]
dynamic = ["version"]

name = "stac-fastapi.static"
description = "An implementation of STAC API based on the FastAPI framework using a static catalog as backend."
keywords = ["STAC", "FastAPI", "stac-fastapi", "static", "fs", "filesystem"]
readme = "README.md"

classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Development Status :: 4 - Beta",
]

authors = [{ name = "Pierre Fontbonne", email = "pierre.fontbonne@uca.fr" }]
maintainers = [
    { name = "Pierre Fontbonne", email = "pierre.fontbonne@uca.fr" },
    { name = "Yannick Guéhenneux", email = "yannick.guehenneux@uca.fr" },
]
license = "etalab-2.0"
license-files = ["LICENCE.txt"]

requires-python = ">= 3.10"

dependencies = [
    "requests>=2.32,<3",
    "requests-cache>=1.2,<2",
    "requests_file>=2.1,<3",
    "xxhash>=3.5,<4",
    "pydantic>=2.10,<3",
    "stac_pydantic>=3.1,<4",
    "stac-fastapi.api>=5.1,<6",
    "stac-fastapi.extensions>=5.1,<6",
    "stac-fastapi.types>=5.1,<6",
    "brotli_asgi>=1.4,<2",
    "shapely>=2.0,<3",
    "orjson>=3.10,<4",
    "cql2>=0.3,<1",
    "uvicorn[standard]",
]

[dependency-groups]

dev = [
    "pytest",
    "pystac",
    "numpy>=1.7,<3",
    "httpx",
    "pystac_client",
    "matplotlib",
]

doc = ["ipykernel>=6.29.5", "notebook>=7.4.3"]

[project.scripts]
stac-fastapi-static = "stac_fastapi.static.app:main"

[project.urls]
Homepage = "https://github.com/fntb/stac-fastapi-static"
Documentation = "https://github.com/fntb/stac-fastapi-static"
Repository = "https://github.com/fntb/stac-fastapi-static"
Issues = "https://github.com/fntb/stac-fastapi-static"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "stac_fastapi.static",
    "stac_fastapi.static.api",
    "stac_fastapi.static.core",
    "stac_fastapi.static.core.compat",
    "stac_fastapi.static.core.lib",
    "stac_fastapi.static.core.model",
    "stac_fastapi.static.core.requests",
]

[tool.setuptools.dynamic]
version = { attr = "stac_fastapi.static.__about__.__version__" }
